+++ Kernel/test.txt
@@ -1,0 +1,7 @@
+@@@@      @     @         @          @@      @
+@   @      @   @         @ @         @ @     @
+@   @       @ @         @   @        @  @    @
+@@@@         @         @     @       @   @   @
+@   @        @        @@@@@@@@@      @    @  @ 
+@    @       @       @         @     @     @ @ 
+@     @      @      @           @    @      @@
